begin reorg
lots of moving stuff around, it will be very broken right now. the plan is to reorganize and then walk through the new workflow to see what's broken and what needs to be changed. I think mostly things just need to be reorganized, not drastically changed in how they function